[Refactor] Split spill restore task from input stream
