+++ be/src/exec/aggregator.h
@@ -32,6 +32,7 @@
 #include "exec/aggregator_fwd.h"
 #include "exec/limited_pipeline_chunk_buffer.h"
 #include "exec/pipeline/context_with_dependency.h"
+#include "exec/pipeline/pipeline_fwd.h"
 #include "exec/pipeline/primitives/pipeline_observer.h"
 #include "exec/pipeline/spill_process_channel.h"
 #include "exprs/agg/aggregate.h"
