+++ be/src/exec/pipeline/aggregate/spillable_aggregate_blocking_source_operator.cpp
@@ -18,6 +18,7 @@
 
 #include "common/status.h"
 #include "exec/pipeline/aggregate/aggregate_blocking_source_operator.h"
+#include "exec/spill/mem_tracker_guard.h"
 
 namespace starrocks::pipeline {
 Status SpillableAggregateBlockingSourceOperator::prepare(RuntimeState* state) {
