+++ be/src/exec/pipeline/aggregate/spillable_aggregate_distinct_blocking_operator.cpp
@@ -19,6 +19,7 @@
 #include "base/concurrency/race_detect.h"
 #include "exec/pipeline/query_context.h"
 #include "exec/sorted_streaming_aggregator.h"
+#include "exec/spill/mem_tracker_guard.h"
 #include "exec/spill/spiller.hpp"
 #include "runtime/runtime_state_helper.h"
