+++ be/src/exec/pipeline/aggregate/spillable_partitionwise_aggregate_operator.cpp
@@ -16,6 +16,8 @@
 #include "exec/pipeline/aggregate/spillable_partitionwise_aggregate_operator.h"
 
 #include "exec/pipeline/query_context.h"
+#include "exec/spill/mem_tracker_guard.h"
+#include "runtime/current_thread.h"
 #include "runtime/runtime_state_helper.h"
 
 namespace starrocks::pipeline {
