+++ be/src/exec/pipeline/exchange/mem_limited_chunk_queue.cpp
@@ -28,9 +28,10 @@
 #include "exec/spill/block_manager.h"
 #include "exec/spill/data_stream.h"
 #include "exec/spill/dir_manager.h"
-#include "exec/spill/executor.h"
 #include "exec/spill/mem_table.h"
+#include "exec/spill/mem_tracker_guard.h"
 #include "exec/spill/options.h"
+#include "exec/spill/task_executor.h"
 #include "fmt/format.h"
 #include "fs/fs.h"
 #include "gen_cpp/InternalService_types.h"
