+++ be/src/exec/pipeline/nljoin/nljoin_context.cpp
@@ -20,8 +20,8 @@
 
 #include "exec/cross_join_node.h"
 #include "exec/pipeline/runtime_filter_hub.h"
-#include "exec/spill/executor.h"
 #include "exec/spill/input_stream.h"
+#include "exec/spill/mem_tracker_guard.h"
 #include "exec/spill/spill_components.h"
 #include "exec/spill/spiller.hpp"
 #include "exprs/expr.h"
