+++ be/src/exec/pipeline/pipeline.cpp
@@ -19,6 +19,7 @@
 #include "exec/pipeline/group_execution/execution_group.h"
 #include "exec/pipeline/operator.h"
 #include "exec/pipeline/pipeline_driver.h"
+#include "exec/pipeline/query_context.h"
 #include "exec/pipeline/scan/connector_scan_operator.h"
 #include "exec/pipeline/scan/schema_scan_operator.h"
 #include "exec/workgroup/pipeline_executor_set.h"
