+++ be/src/exec/pipeline/sort/partition_sort_sink_operator.h
@@ -25,7 +25,6 @@
 #include "exec/pipeline/sort/sort_context.h"
 #include "exec/pipeline/spill_process_channel.h"
 #include "exec/sort_exec_exprs.h"
-#include "exec/spill/executor.h"
 #include "exec/spill/spiller_factory.h"
 #include "runtime/runtime_state_fwd.h"
