+++ be/src/exec/spill/spill_components.h
@@ -21,14 +21,15 @@
 #include "base/concurrency/race_detect.h"
 #include "column/vectorized_fwd.h"
 #include "common/status.h"
+#include "exec/spill/block_group.h"
 #include "exec/spill/block_manager.h"
 #include "exec/spill/data_stream.h"
-#include "exec/spill/executor.h"
 #include "exec/spill/input_stream.h"
 #include "exec/spill/mem_table.h"
 #include "exec/spill/options.h"
 #include "exec/spill/partition.h"
 #include "exec/spill/serde.h"
+#include "exec/spill/task_executor.h"
 #include "exec/workgroup/scan_task.h"
 #include "fmt/format.h"
 #include "fs/fs.h"
