+++ be/src/exec/spill/spiller.h
@@ -28,7 +28,6 @@
 #include "common/status.h"
 #include "exec/spill/block_manager.h"
 #include "exec/spill/common.h"
-#include "exec/spill/executor.h"
 #include "exec/spill/input_stream.h"
 #include "exec/spill/mem_table.h"
 #include "exec/spill/options.h"
@@ -37,6 +36,7 @@
 #include "exec/spill/spill_components.h"
 #include "exec/spill/spill_metrics.h"
 #include "exec/spill/spiller_factory.h"
+#include "exec/spill/task_executor.h"
 #include "fs/fs.h"
 #include "gen_cpp/InternalService_types.h"
 #include "runtime/mem_tracker.h"
