+++ be/src/exec/spill/spiller.hpp
@@ -28,11 +28,14 @@
 #include "common/runtime_profile.h"
 #include "common/status.h"
 #include "exec/spill/common.h"
-#include "exec/spill/executor.h"
 #include "exec/spill/input_stream.h"
+#include "exec/spill/mem_tracker_guard.h"
+#include "exec/spill/restore_task.h"
+#include "exec/spill/runtime_context.h"
 #include "exec/spill/serde.h"
 #include "exec/spill/spill_components.h"
 #include "exec/spill/spiller.h"
+#include "exec/spill/task_executor.h"
 #include "exec/workgroup/work_group_fwd.h"
 #include "gen_cpp/InternalService_types.h"
 #include "runtime/runtime_state_fwd.h"
