+++ be/src/runtime/load_channel.cpp
@@ -47,6 +47,7 @@
 #include "common/util/thrift_util.h"
 #include "fmt/format.h"
 #include "runtime/closure_guard.h"
+#include "runtime/current_thread.h"
 #include "runtime/descriptors.h"
 #include "runtime/diagnose_daemon.h"
 #include "runtime/lake_tablets_channel.h"
