+++ be/src/storage/lake/tablet_internal_parallel_merge_task.cpp
@@ -18,11 +18,12 @@
 #include "column/chunk_schema_helper.h"
 #include "common/config_exec_fwd.h"
 #include "common/runtime_profile.h"
-#include "exec/spill/input_stream.h"
+#include "exec/spill/block_group.h"
 #include "exec/spill/options.h"
 #include "exec/spill/serde.h"
 #include "exec/spill/spiller.h"
 #include "exec/spill/spiller_factory.h"
+#include "runtime/current_thread.h"
 #include "runtime/exec_env.h"
 #include "runtime/runtime_state.h"
 #include "storage/aggregate_iterator.h"
