+++ be/src/storage/load_spill_block_manager.h
@@ -17,9 +17,9 @@
 #include <utility>
 
 #include "common/thread/threadpool.h"
+#include "exec/spill/block_group.h"
 #include "exec/spill/block_manager.h"
 #include "exec/spill/dir_manager.h"
-#include "exec/spill/input_stream.h"
 #include "fmt/format.h"
 #include "storage/lake/location_provider.h"
