+++ be/test/io/spill_test.cpp
@@ -44,13 +44,14 @@
 #include "common/statusor.h"
 #include "compute_env/sorting/merge.h"
 #include "compute_env/sorting/sorting.h"
-#include "exec/spill/executor.h"
 #include "exec/spill/log_block_manager.h"
 #include "exec/spill/mem_table.h"
+#include "exec/spill/mem_tracker_guard.h"
 #include "exec/spill/spill_components.h"
 #include "exec/spill/spiller.h"
 #include "exec/spill/spiller.hpp"
 #include "exec/spill/spiller_factory.h"
+#include "exec/spill/task_executor.h"
 #include "exec/workgroup/scan_task.h"
 #include "exprs/column_ref.h"
 #include "exprs/expr_context.h"
